Add Goal Quick Function

+++ database/migrations/2025_12_16_072319_contributions.php
@@ -12,7 +12,7 @@ class Contributions extends Migration {
 
             $table->increments('contribution_id');
 
-            $table->integer('player_id')->unsigned();
+            $table->bigInteger('player_id')->unsigned();
             $table->foreign('player_id')->references('id')->on('players');
 
             $table->double('amount')->default(0);
